+++ Swift/Example/Example.xcodeproj/xcuserdata/kruperfone.xcuserdatad/xcschemes/Example.xcscheme
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "210C52AC1BFA3C8100F6C390"
-               BuildableName = "FSHelpers.framework"
-               BlueprintName = "FSHelpers"
-               ReferencedContainer = "container:FSHelpers.xcodeproj">
+               BlueprintIdentifier = "210C527F1BFA3C6400F6C390"
+               BuildableName = "Example.app"
+               BlueprintName = "Example"
+               ReferencedContainer = "container:Example.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -32,20 +32,20 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "210C52B61BFA3C8100F6C390"
-               BuildableName = "FSHelpersTests.xctest"
-               BlueprintName = "FSHelpersTests"
-               ReferencedContainer = "container:FSHelpers.xcodeproj">
+               BlueprintIdentifier = "210C52931BFA3C6500F6C390"
+               BuildableName = "ExampleTests.xctest"
+               BlueprintName = "ExampleTests"
+               ReferencedContainer = "container:Example.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "210C52AC1BFA3C8100F6C390"
-            BuildableName = "FSHelpers.framework"
-            BlueprintName = "FSHelpers"
-            ReferencedContainer = "container:FSHelpers.xcodeproj">
+            BlueprintIdentifier = "210C527F1BFA3C6400F6C390"
+            BuildableName = "Example.app"
+            BlueprintName = "Example"
+            ReferencedContainer = "container:Example.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -61,15 +61,16 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <MacroExpansion>
+      <BuildableProductRunnable
+         runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "210C52AC1BFA3C8100F6C390"
-            BuildableName = "FSHelpers.framework"
-            BlueprintName = "FSHelpers"
-            ReferencedContainer = "container:FSHelpers.xcodeproj">
+            BlueprintIdentifier = "210C527F1BFA3C6400F6C390"
+            BuildableName = "Example.app"
+            BlueprintName = "Example"
+            ReferencedContainer = "container:Example.xcodeproj">
          </BuildableReference>
-      </MacroExpansion>
+      </BuildableProductRunnable>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
@@ -79,15 +80,16 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <MacroExpansion>
+      <BuildableProductRunnable
+         runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "210C52AC1BFA3C8100F6C390"
-            BuildableName = "FSHelpers.framework"
-            BlueprintName = "FSHelpers"
-            ReferencedContainer = "container:FSHelpers.xcodeproj">
+            BlueprintIdentifier = "210C527F1BFA3C6400F6C390"
+            BuildableName = "Example.app"
+            BlueprintName = "Example"
+            ReferencedContainer = "container:Example.xcodeproj">
          </BuildableReference>
-      </MacroExpansion>
+      </BuildableProductRunnable>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
