+++ Swift/Example/Example/Assets.xcassets/AppIcon.appiconset/Contents.json
