+++ Swift/Example/Example/ViewController.swift
@@ -1,0 +1,27 @@
+//
+//  ViewController.swift
+//  Example
+//
+//  Created by Kruperfone on 16.11.15.
+//  Copyright © 2015 FlatStack. All rights reserved.
+//
+
+import UIKit
+import FSHelpers
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+
+
+}
+
