+++ Swift/FSHelpers/FSHelpers/FSHelpers.h
@@ -3,7 +3,7 @@
 //  FSHelpers
 //
 //  Created by Kruperfone on 16.11.15.
-//
+//  Copyright © 2015 FlatStack. All rights reserved.
 //
 
 #import <UIKit/UIKit.h>
