+++ Swift/Helpers/FSExtensions/FSE+Double.swift
@@ -1,0 +1,17 @@
+//
+//  FSE+Double.swift
+//  SwiftHelpers
+//
+//  Created by Kruperfone on 31.07.15.
+//  Copyright (c) 2015 FlatStack. All rights reserved.
+//
+
+import Foundation
+
+public extension Double {
+    public var fs_dispatchTime: dispatch_time_t {
+        get {
+            return dispatch_time(DISPATCH_TIME_NOW,Int64(self * Double(NSEC_PER_SEC)))
+        }
+    }
+}
