+++ Swift/HelpersWorkspace.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "group:Example/Example.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:FSHelpers/FSHelpers.xcodeproj">
+   </FileRef>
+</Workspace>
