--- Swift/SwiftHelpers/SwiftHelpers.xcodeproj/project.xcworkspace/xcshareddata/SwiftHelpers.xccheckout
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
-	<false/>
-	<key>IDESourceControlProjectIdentifier</key>
-	<string>4232A81D-2D22-40C2-95E0-C86A8830F92C</string>
-	<key>IDESourceControlProjectName</key>
-	<string>SwiftHelpers</string>
-	<key>IDESourceControlProjectOriginsDictionary</key>
-	<dict>
-		<key>5498BAFDC3AD6CE5D14E6E181524F43733D41CD0</key>
-		<string>github.com:fs/FSHelper.git</string>
-	</dict>
-	<key>IDESourceControlProjectPath</key>
-	<string>Swift/Project/SwiftHelpers.xcodeproj</string>
-	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
-	<dict>
-		<key>5498BAFDC3AD6CE5D14E6E181524F43733D41CD0</key>
-		<string>../../../..</string>
-	</dict>
-	<key>IDESourceControlProjectURL</key>
-	<string>github.com:fs/FSHelper.git</string>
-	<key>IDESourceControlProjectVersion</key>
-	<integer>111</integer>
-	<key>IDESourceControlProjectWCCIdentifier</key>
-	<string>5498BAFDC3AD6CE5D14E6E181524F43733D41CD0</string>
-	<key>IDESourceControlProjectWCConfigurations</key>
-	<array>
-		<dict>
-			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
-			<string>public.vcs.git</string>
-			<key>IDESourceControlWCCIdentifierKey</key>
-			<string>5498BAFDC3AD6CE5D14E6E181524F43733D41CD0</string>
-			<key>IDESourceControlWCCName</key>
-			<string>FSHelper</string>
-		</dict>
-	</array>
-</dict>
-</plist>
