--- Swift/SwiftHelpers/SwiftHelpers.xcodeproj/xcuserdata/kruperfone.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>SwiftHelpers.xcscheme</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-		<key>Widget.xcscheme</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>1</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>21D98A0B1B6B76860009E7D3</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>21D98A201B6B76860009E7D3</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>21D98A341B6B76C70009E7D3</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
